Use points for actual efficiency values
